Add mouse hovered event; Create test GUI buttons; Update version to 0.4.5; Clean up code; Tweak project build settings.

+++ src/libgame/util/combine.hpp
@@ -17,7 +17,7 @@ std::string combineStringPath(First firstString, Rest&... rest)
 template <typename First>
 int combineModuleTypes(First firstModule = 0)
 {
-	return firstModule;
+	return (int) firstModule;
 }
 
 template <typename First, typename ... Rest>
